+++ python/dgl/node_flow.py
@@ -10,126 +10,7 @@ from .frame import Frame, FrameRef
 from .graph import DGLBaseGraph
 from .runtime import ir, scheduler, Runtime
 from . import utils
-
-NodeSpace = namedtuple('NodeSpace', ['data'])
-
-class LayerView(object):
-    """A NodeView class to act as G.nodes for a DGLGraph.
-
-    Can be used to get a list of current nodes and get and set node data.
-
-    See Also
-    --------
-    dgl.DGLGraph.nodes
-    """
-    __slots__ = ['_graph']
-
-    def __init__(self, graph):
-        self._graph = graph
-
-    def __len__(self):
-        return self._graph.num_layers()
-
-    def __getitem__(self, layer):
-        if not isinstance(layer, int):
-            raise DGLError('Currently we only support the view of one layer')
-        return NodeSpace(data=LayerDataView(self._graph, layer))
-
-    def __call__(self):
-        """Return the nodes."""
-        return F.arange(0, len(self))
-
-class LayerDataView(MutableMapping):
-    """The data view class when G.nodes[...].data is called.
-
-    See Also
-    --------
-    dgl.DGLGraph.nodes
-    """
-    __slots__ = ['_graph', '_layer']
-
-    def __init__(self, graph, layer):
-        self._graph = graph
-        self._layer = layer
-
-    def __getitem__(self, key):
-        return self._graph._node_frames[self._layer][key]
-
-    def __setitem__(self, key, val):
-        self._graph._node_frames[self._layer][key] = val
-
-    def __delitem__(self, key):
-        del self._graph._node_frames[self._layer][key]
-
-    def __len__(self):
-        return len(self._graph._node_frames[self._layer])
-
-    def __iter__(self):
-        return iter(self._graph._node_frames[self._layer])
-
-    def __repr__(self):
-        data = self._graph._node_frames[self._layer]
-        return repr({key : data[key] for key in data})
-
-EdgeSpace = namedtuple('EdgeSpace', ['data'])
-
-class BlockView(object):
-    """A EdgeView class to act as G.edges for a DGLGraph.
-
-    Can be used to get a list of current edges and get and set edge data.
-
-    See Also
-    --------
-    dgl.DGLGraph.edges
-    """
-    __slots__ = ['_graph']
-
-    def __init__(self, graph):
-        self._graph = graph
-
-    def __len__(self):
-        return self._graph.num_blocks
-
-    def __getitem__(self, flow):
-        if not isinstance(flow, int):
-            raise DGLError('Currently we only support the view of one flow')
-        return EdgeSpace(data=BlockDataView(self._graph, flow))
-
-    def __call__(self, *args, **kwargs):
-        """Return all the edges."""
-        return self._graph.all_edges(*args, **kwargs)
-
-class BlockDataView(MutableMapping):
-    """The data view class when G.edges[...].data is called.
-
-    See Also
-    --------
-    dgl.DGLGraph.edges
-    """
-    __slots__ = ['_graph', '_flow']
-
-    def __init__(self, graph, flow):
-        self._graph = graph
-        self._flow = flow
-
-    def __getitem__(self, key):
-        return self._graph._edge_frames[self._flow][key]
-
-    def __setitem__(self, key, val):
-        self._graph._edge_frames[self._flow][key] = val
-
-    def __delitem__(self, key):
-        del self._graph._edge_frames[self._flow][key]
-
-    def __len__(self):
-        return len(self._graph._edge_frames[self._flow])
-
-    def __iter__(self):
-        return iter(self._graph._edge_frames[self._flow])
-
-    def __repr__(self):
-        data = self._graph._edge_frames[self._flow]
-        return repr({key : data[key] for key in data})
+from .view import LayerView, BlockView
 
 def _copy_to_like(arr1, arr2):
     return F.copy_to(arr1, F.context(arr2))
